+++ stroom-statistics/stroom-statistics-impl-sql/src/main/java/stroom/statistics/impl/sql/search/StatStoreCriteriaBuilder.java
@@ -9,6 +9,7 @@ import stroom.statistics.impl.sql.shared.StatisticRollUpType;
 import stroom.statistics.impl.sql.shared.StatisticStoreDoc;
 import stroom.util.Period;
 import stroom.util.logging.LogUtil;
+import stroom.util.rest.RestUtil;
 import stroom.util.shared.Range;
 
 import org.slf4j.Logger;
@@ -20,7 +21,6 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import javax.ws.rs.BadRequestException;
 
 public class StatStoreCriteriaBuilder {
 
@@ -47,7 +47,7 @@ public class StatStoreCriteriaBuilder {
 //        topLevelExpressionOperator = ExpressionUtil.replaceExpressionParameters(topLevelExpressionOperator, paramMap);
 
         if (expression == null || expression.op() == null) {
-            throw new BadRequestException(
+            throw RestUtil.badRequest(
                     "The top level operator for the query must be one of [" + ExpressionOperator.Op.values() + "]");
         }
 
@@ -71,7 +71,7 @@ public class StatStoreCriteriaBuilder {
                         final ExpressionTerm expressionTerm = (ExpressionTerm) expressionItem;
 
                         if (expressionTerm.getField() == null) {
-                            throw new BadRequestException("Expression term does not have a field specified");
+                            throw RestUtil.badRequest("Expression term does not have a field specified");
                         }
 
                         if (expressionTerm.getField().equals(StatisticStoreDoc.FIELD_NAME_DATE_TIME)) {
@@ -84,7 +84,7 @@ public class StatStoreCriteriaBuilder {
                         }
                     } else if (expressionItem instanceof ExpressionOperator) {
                         if (((ExpressionOperator) expressionItem).op() == null) {
-                            throw new BadRequestException("An operator in the query is missing a type, it should " +
+                            throw RestUtil.badRequest("An operator in the query is missing a type, it should " +
                                     "be one of " + ExpressionOperator.Op.values());
                         }
                     }
@@ -94,7 +94,7 @@ public class StatStoreCriteriaBuilder {
 
         // ensure we have a date term
         if (dateTermsFound != 1 || validDateTermsFound != 1) {
-            throw new BadRequestException(
+            throw RestUtil.badRequest(
                     "Search queries on the statistic store must contain one term using the '"
                             + StatisticStoreDoc.FIELD_NAME_DATE_TIME
                             + "' field with one of the following condtitions [" + SUPPORTED_DATE_CONDITIONS.toString()
@@ -103,7 +103,7 @@ public class StatStoreCriteriaBuilder {
 
         // ensure the value field is not used in the query terms
         if (contains(expression, StatisticStoreDoc.FIELD_NAME_VALUE)) {
-            throw new BadRequestException("Search queries containing the field '"
+            throw RestUtil.badRequest("Search queries containing the field '"
                     + StatisticStoreDoc.FIELD_NAME_VALUE + "' are not supported.  Please remove it from the query");
         }
 
@@ -130,11 +130,11 @@ public class StatStoreCriteriaBuilder {
 
         if (!rolledUpFieldNames.isEmpty()) {
             if (dataSource.getRollUpType().equals(StatisticRollUpType.NONE)) {
-                throw new BadRequestException(
+                throw RestUtil.badRequest(
                         "Query contains rolled up terms but the Statistic Data Source does not support any roll-ups");
             } else if (dataSource.getRollUpType().equals(StatisticRollUpType.CUSTOM)) {
                 if (!dataSource.isRollUpCombinationSupported(rolledUpFieldNames)) {
-                    throw new BadRequestException(String.format("The query contains a combination of rolled up " +
+                    throw RestUtil.badRequest(String.format("The query contains a combination of rolled up " +
                             "fields %s that is not in the list of custom roll-ups for the statistic data " +
                             "source", rolledUpFieldNames));
                 }
@@ -175,7 +175,7 @@ public class StatStoreCriteriaBuilder {
             } else if (node instanceof ExpressionOperator) {
                 ExpressionOperator expressionOperator = (ExpressionOperator) node;
                 if (expressionOperator.getChildren() == null) {
-                    throw new BadRequestException(LogUtil.message("{} operator contains no terms or operators",
+                    throw RestUtil.badRequest(LogUtil.message("{} operator contains no terms or operators",
                             expressionOperator.op().getDisplayValue()));
                 }
                 for (final ExpressionItem childNode : expressionOperator.getChildren()) {
@@ -215,7 +215,7 @@ public class StatStoreCriteriaBuilder {
         final String[] dateArr = dateTerm.getValue().split(",");
 
         if (dateArr.length != 2) {
-            throw new BadRequestException("DateTime term is not a valid format, term: " + dateTerm.toString());
+            throw RestUtil.badRequest("DateTime term is not a valid format, term: " + dateTerm.toString());
         }
 
         rangeFrom = parseDateTime("from", dateArr[0], timeZoneId, nowEpochMilli);
@@ -234,10 +234,10 @@ public class StatStoreCriteriaBuilder {
         final ZonedDateTime dateTime;
         try {
             dateTime = DateExpressionParser.parse(value, timeZoneId, nowEpochMilli)
-                    .orElseThrow(() -> new BadRequestException(
+                    .orElseThrow(() -> RestUtil.badRequest(
                             "DateTime term has an invalid '" + type + "' value of '" + value + "'"));
         } catch (final Exception e) {
-            throw new BadRequestException("DateTime term has an invalid '" + type + "' value of '" + value + "'");
+            throw RestUtil.badRequest("DateTime term has an invalid '" + type + "' value of '" + value + "'");
         }
 
         return dateTime.toInstant().toEpochMilli();